+++ meta/template/basic.lua
@@ -210,7 +210,7 @@ function select(index, ...) end
 
 ---@version 5.1
 ---#DES 'setfenv'
----@param f     async fun(...):...|integer
+---@param f     (async fun(...):...)|integer
 ---@param table table
 ---@return function
 function setfenv(f, table) end