+++ changelog.md
@@ -4,6 +4,7 @@
 <!-- Add all new changes here. They will be moved under a version at release -->
 * `NEW` Custom documentation exporter
 * `NEW` Setting: `Lua.docScriptPath`: Path to a script that overrides `cli.doc.export`, allowing user-specified documentation exporting.
+* `FIX` Fix `VM.OnCompileFunctionParam` function in plugins
 
 ## 3.10.5
 `2024-8-19`